cloned_repo/requirements.txt
langgraph
langchain
langchain-core
langchain-openai
langchain-community

openai
pydantic
python-dotenv

streamlit
pandas

tavily-python

google-genai

typing-extensions
requests
aiohttp
asyncio
nest-asyncio
protobuf

python-dotenv